Fix a for loop in main.lua (#85)
* Fix a for loop in main.lua

For loop missing the extra variable. Though I haven't used ESX in a while I am assuming there are only 4 windows for each door. Though I also adjusted it to meet the fivem native of window ids which start at 0 and go to 3 to indicate each window on the doors. 

If this is wrong or the ids are wrong then i can correct them.

* Update main.lua